Splited the code into components
